mc: bloom_filter: regroup hash function in misc/hashfunc
* spot/mc/bitstate.hh,
  spot/mc/mc.hh,
  tests/ltsmin/modelcheck.cc: use size_t for mem_size
* spot/mc/bloom_filter.hh,
  spot/misc/hashfunc.hh: regroup hash function

+++ spot/mc/bitstate.hh
@@ -29,7 +29,7 @@ namespace spot
   class bitstate
   {
   public:
-    bitstate(kripkecube<State, SuccIterator>& sys, uint32_t mem_size):
+    bitstate(kripkecube<State, SuccIterator>& sys, size_t mem_size):
       sys_(sys)
     {
       static_assert(spot::is_a_kripkecube_ptr<decltype(&sys),
